Decrease school and workplace sizes
In the prior version, setting parameters to get Ro in the range
of 2-3 had quick outbreaks that infected <70% of the population.
Likely cause was large school sizes (400-800) and some large
workplaces. Revised so that schools are size 100 students and
that workplaces are 100 or fewer workers. Made corresponding
changes to the Base_Model and Estimate_R0 code.