Change model parameters: drop scale, add mask switch
First attempts at model fitting suggest that the scale parameter
(which reduced the effect of mobility measures on mobility) is
not necessary, so it was removed. But later transmission does not
seem to be explainable only by mobility, so added a "masks"
parameter to slightly reduce transmission in community settings.